+++ modules/agent/src/main/java/io/jpom/common/AgentExceptionHandler.java
@@ -22,6 +22,7 @@
  */
 package io.jpom.common;
 
+import cn.hutool.core.exceptions.ValidateException;
 import cn.hutool.extra.servlet.ServletUtil;
 import cn.jiangzeyin.common.DefaultSystemLog;
 import cn.jiangzeyin.common.JsonMessage;
@@ -66,7 +67,7 @@ public class AgentExceptionHandler {
 	 * @param response 响应
 	 * @param e        异常
 	 */
-	@ExceptionHandler({IllegalArgumentException.class, IllegalStateException.class})
+	@ExceptionHandler({IllegalArgumentException.class, IllegalStateException.class, ValidateException.class})
 	public void paramExceptionHandler(HttpServletRequest request, HttpServletResponse response, Exception e) {
 		DefaultSystemLog.getLog().error("controller " + request.getRequestURI(), e);
 		ServletUtil.write(response, JsonMessage.getString(405, e.getMessage()), MediaType.APPLICATION_JSON_VALUE);